checks daily to see if current date is in slated term dates
