+++ app/server.js
@@ -4,68 +4,11 @@ import moment from 'moment'
 import { createUser, getDALIUsers } from './db-actions/user-actions'
 import { pokeChannels, getPokeChannels } from './data-actions/channel-productivity'
 import { checkChannelActivity } from './data-actions/automate-tasks'
-import { getMilestone } from './data-actions/milestones'
+import { getMilestone, checkOnTerm } from './data-actions/milestones'
 
 let currentWeek = 0
 let onTerm = false
-
-/*
- * Shifts the provided date to the first Wednesday
- */
-const shiftToWednesday = (termStartDate) => {
-  while (termStartDate.format('dddd') !== 'Wednesday') {
-    termStartDate.add(1, 'days')
-  }
-  return termStartDate
-}
-
-// winter
-const firstWeekWinter = moment().week(1)
-let firstDayWinter = firstWeekWinter.add(3 - firstWeekWinter.get('date'), 'days')
-let lastDayWinter = moment()
-while (firstDayWinter.format('dddd') !== 'Monday' && firstDayWinter.format('dddd') !== 'Wednesday') {
-  firstDayWinter.add(1, 'days')
-}
-firstDayWinter = shiftToWednesday(firstDayWinter)
-lastDayWinter = firstDayWinter.add(10, 'weeks')
-console.log(`First day of winter term: ${firstDayWinter}`)
-console.log(`Last day of winter term: ${lastDayWinter}`)
-
-// spring
-const firstWeekSpring = moment().week(12)
-let firstDaySpring = firstWeekSpring
-let lastDaySpring = moment()
-while (firstDaySpring.format('dddd') !== 'Monday') {
-  firstDaySpring.add(1, 'days')
-}
-firstDaySpring = shiftToWednesday(firstDaySpring)
-lastDaySpring = firstDaySpring.add(10, 'weeks')
-console.log(`First day of spring term: ${firstDaySpring}`)
-console.log(`Last day of winter term: ${lastDaySpring}`)
-
-// summer
-const firstWeekSummer = moment().week(24)
-let firstDaySummer = firstWeekSummer
-let lastDaySummer = moment()
-while (firstDaySummer.format('dddd') !== 'Thursday') {
-  firstDaySummer.add(1, 'days')
-}
-firstDaySummer = shiftToWednesday(firstDaySummer)
-lastDaySummer = firstDaySummer.add(10, 'weeks')
-console.log(`First day of summer term: ${firstDaySummer}`)
-console.log(`Last day of summer term: ${lastDaySummer}`)
-
-// fall
-const firstWeekFall = moment().week(36)
-let firstDayFall = firstWeekFall
-let lastDayFall = moment()
-while ((firstDayFall.format('dddd') !== 'Monday' && firstDayFall.format('dddd') !== 'Wednesday') || firstDayFall.get('date') < 11) {
-  firstDayFall.add(1, 'days')
-}
-firstDayFall = shiftToWednesday(firstDayFall)
-lastDayFall = firstDayFall.add(10, 'weeks')
-console.log(`First day of summer term: ${firstDayFall}`)
-console.log(`Last day of summer term: ${lastDayFall}`)
+let checkTermBoundaries = false
 
 // botkit controller
 const controller = botkit.slackbot({
@@ -216,4 +159,20 @@ const milestoneReminder = schedule.scheduleJob({ hour: 10, minute: 0, second: 0,
     // reset the week counter
     if (currentWeek === 10) currentWeek = 0
   })
+})
+
+// Checks daily at 12AM to see if term start/end dates are correct
+const updateTermBounds = schedule.scheduleJob({ hour: 0, minute: 0, second: 0 }, () => {
+  console.log('Updating the expected term bounds are correct')
+
+  // We are not in a term and we haven't confirmed correct term start/end dates
+  if (!checkTermBoundaries && !onTerm) {
+    /*
+     * TODO: Check with admin to make sure that currently assigned
+     * start/end dates are correct
+     */
+    onTerm = checkOnTerm()
+    if (onTerm) checkTermBoundaries = false
+    else checkTermBoundaries = true
+  }
 })
